requirements.txt
django>=4.2
playwright
pillow
python-dotenv
replicate
requests